requirements.txt
litellm>=1.40.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
httpx>=0.27.0
streamlit>=1.35.0
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
bleach>=6.0.0
